1.add result map 2.update proxserver to adminserver
