+++ app/dao/db/index.js
@@ -79,11 +79,11 @@ databases.forEach((database) => {
 	let tableObj = {};
 	let dbModelsPath = __dirname + '/' + database + '_models';
 	let dbModels = fs.readdirSync(dbModelsPath);
+	let alter = database == "db_tars_web"
 	dbModels.forEach(function (dbModel) {
 		let tableName = dbModel.replace(/\.js$/g, '');
 		tableObj[_.camelCase(tableName)] = sequelize.import(dbModelsPath + '/' + tableName);
-		// tableObj[_.camelCase(tableName)].sync({ alter: true });
-		tableObj[_.camelCase(tableName)].sync();
+		tableObj[_.camelCase(tableName)].sync({ alter: alter });
 	});
 	Db[database] = tableObj;
 	Db[database].sequelize = sequelize;
