+++ app/service/util/rpcClient/index.js
@@ -19,7 +19,7 @@ const AdminRegProxy = require("./rpcProxy/AdminRegProxy");
 const ConfigFProxy = require("./rpcProxy/ConfigFProxy");
 const DCacheOptProxy = require("./rpcProxy/DCacheOptProxy");
 const MonitorQueryProxy = require("./rpcProxy/MonitorQueryProxy");
-const ProxyProxy = require("./rpcProxy/ProxyProxy");
+const BenchmarkAdminProxy = require("./rpcProxy/BenchmarkAdminProxy");
 const path = require('path');
 const logger = require('./../../../logger');
 client.initialize(path.join(__dirname, '../../../../config/tars.conf'));
@@ -104,8 +104,8 @@ module.exports = {
 
     // nodePrx: RPCClientPrx(NodeProxy, 'tars', 'Node', 'tars.tarsnode.NodeObj'),
     // nodeStruct: RPCStruct(NodeProxy, 'tars'),
-    benchmarkPrx : RPCClientPrx(ProxyProxy, 'bm', 'Proxy', 'bm.ProxyServer.ProxyObj'),
-    benchmarkStruct : RPCStruct(ProxyProxy, 'bm'),
+    benchmarkPrx : RPCClientPrx(BenchmarkAdminProxy, 'bm', 'Admin', 'bm.AdminServer.AdminObj'),
+    benchmarkStruct : RPCStruct(BenchmarkAdminProxy, 'bm'),
 
     client: client
 };
