+++ app/service/util/rpcClient/rpcProxy/BenchmarkAdminProxy.js
@@ -1,0 +1,591 @@
+// **********************************************************************
+// Parsed By TarsParser(1.1.0), Generated By tars2node(20190529)
+// TarsParser Maintained By <TARS> and tars2node Maintained By <superzheng>
+// Generated from "BenchmarkAdmin.tars" by Client Mode
+// **********************************************************************
+
+/* eslint-disable */
+
+"use strict";
+
+var assert    = require("assert");
+var TarsStream = require("@tars/stream");
+var TarsError  = require("@tars/rpc").error;
+var _TARS_MODULE_A_ = require("./BenchmarkNodeTars.js");
+
+var _hasOwnProperty = Object.prototype.hasOwnProperty;
+var _makeError = function (data, message, type) {
+    var error = new Error(message || "");
+    error.request = data.request;
+    error.response = {
+        "costtime" : data.request.costtime
+    };
+    if (type === TarsError.CLIENT.DECODE_ERROR) {
+        error.name = "DECODE_ERROR";
+        error.response.error = {
+            "code" : type,
+            "message" : message
+        };
+    } else {
+        error.name = "RPC_ERROR";
+        error.response.error = data.error;
+    }
+    return error;
+};
+
+var bm = bm || {};
+module.exports.bm = bm;
+
+bm.AdminProxy = function () {
+    this._name    = undefined;
+    this._worker  = undefined;
+};
+
+bm.AdminProxy.prototype.setTimeout = function (iTimeout) {
+    this._worker.timeout = iTimeout;
+};
+
+bm.AdminProxy.prototype.getTimeout = function () {
+    return this._worker.timeout;
+};
+
+bm.AdminProxy.prototype.setVersion = function (iVersion) {
+    this._worker.version = iVersion;
+};
+
+bm.AdminProxy.prototype.getVersion = function () {
+    return this._worker.version;
+};
+
+bm.TaskStat = function() {
+    this.conf = new _TARS_MODULE_A_.bm.TaskConf;
+    this.state = 0;
+    this.start_time = 0;
+    this.fetch_time = 0;
+    this.duration = 0;
+    this.speed_quota = new TarsStream.Map(TarsStream.String, TarsStream.Int32);
+    this._classname = "bm.TaskStat";
+};
+bm.TaskStat._classname = "bm.TaskStat";
+bm.TaskStat._write = function (os, tag, value) { os.writeStruct(tag, value); };
+bm.TaskStat._read  = function (is, tag, def) { return is.readStruct(tag, true, def); };
+bm.TaskStat._readFrom = function (is) {
+    var tmp = new bm.TaskStat;
+    tmp.conf = is.readStruct(0, false, _TARS_MODULE_A_.bm.TaskConf);
+    tmp.state = is.readInt32(1, false, 0);
+    tmp.start_time = is.readInt64(2, false, 0);
+    tmp.fetch_time = is.readInt64(3, false, 0);
+    tmp.duration = is.readInt64(4, false, 0);
+    tmp.speed_quota = is.readMap(5, false, TarsStream.Map(TarsStream.String, TarsStream.Int32));
+    return tmp;
+};
+bm.TaskStat.prototype._writeTo = function (os) {
+    os.writeStruct(0, this.conf);
+    os.writeInt32(1, this.state);
+    os.writeInt64(2, this.start_time);
+    os.writeInt64(3, this.fetch_time);
+    os.writeInt64(4, this.duration);
+    os.writeMap(5, this.speed_quota);
+};
+bm.TaskStat.prototype._equal = function () {
+    assert.fail("this structure not define key operation");
+};
+bm.TaskStat.prototype._genKey = function () {
+    if (!this._proto_struct_name_) {
+        this._proto_struct_name_ = "STRUCT" + Math.random();
+    }
+    return this._proto_struct_name_;
+};
+bm.TaskStat.prototype.toObject = function() { 
+    return {
+        "conf" : this.conf.toObject(),
+        "state" : this.state,
+        "start_time" : this.start_time,
+        "fetch_time" : this.fetch_time,
+        "duration" : this.duration,
+        "speed_quota" : this.speed_quota.toObject()
+    };
+};
+bm.TaskStat.prototype.readFromObject = function(json) { 
+    _hasOwnProperty.call(json, "conf") && (this.conf.readFromObject(json.conf));
+    _hasOwnProperty.call(json, "state") && (this.state = json.state);
+    _hasOwnProperty.call(json, "start_time") && (this.start_time = json.start_time);
+    _hasOwnProperty.call(json, "fetch_time") && (this.fetch_time = json.fetch_time);
+    _hasOwnProperty.call(json, "duration") && (this.duration = json.duration);
+    _hasOwnProperty.call(json, "speed_quota") && (this.speed_quota.readFromObject(json.speed_quota));
+    return this;
+};
+bm.TaskStat.prototype.toBinBuffer = function () {
+    var os = new TarsStream.TarsOutputStream();
+    this._writeTo(os);
+    return os.getBinBuffer();
+};
+bm.TaskStat.new = function () {
+    return new bm.TaskStat();
+};
+bm.TaskStat.create = function (is) {
+    return bm.TaskStat._readFrom(is);
+};
+
+bm.BenchmarkSummary = function() {
+    this.nodes = new TarsStream.Map(TarsStream.String, _TARS_MODULE_A_.bm.NodeStat);
+    this.result = new TarsStream.Map(TarsStream.String, _TARS_MODULE_A_.bm.ResultStat);
+    this.total_result = new TarsStream.Map(TarsStream.String, _TARS_MODULE_A_.bm.ResultStat);
+    this.task = new TarsStream.Map(TarsStream.String, bm.TaskStat);
+    this._classname = "bm.BenchmarkSummary";
+};
+bm.BenchmarkSummary._classname = "bm.BenchmarkSummary";
+bm.BenchmarkSummary._write = function (os, tag, value) { os.writeStruct(tag, value); };
+bm.BenchmarkSummary._read  = function (is, tag, def) { return is.readStruct(tag, true, def); };
+bm.BenchmarkSummary._readFrom = function (is) {
+    var tmp = new bm.BenchmarkSummary;
+    tmp.nodes = is.readMap(0, false, TarsStream.Map(TarsStream.String, _TARS_MODULE_A_.bm.NodeStat));
+    tmp.result = is.readMap(1, false, TarsStream.Map(TarsStream.String, _TARS_MODULE_A_.bm.ResultStat));
+    tmp.total_result = is.readMap(2, false, TarsStream.Map(TarsStream.String, _TARS_MODULE_A_.bm.ResultStat));
+    tmp.task = is.readMap(3, false, TarsStream.Map(TarsStream.String, bm.TaskStat));
+    return tmp;
+};
+bm.BenchmarkSummary.prototype._writeTo = function (os) {
+    os.writeMap(0, this.nodes);
+    os.writeMap(1, this.result);
+    os.writeMap(2, this.total_result);
+    os.writeMap(3, this.task);
+};
+bm.BenchmarkSummary.prototype._equal = function () {
+    assert.fail("this structure not define key operation");
+};
+bm.BenchmarkSummary.prototype._genKey = function () {
+    if (!this._proto_struct_name_) {
+        this._proto_struct_name_ = "STRUCT" + Math.random();
+    }
+    return this._proto_struct_name_;
+};
+bm.BenchmarkSummary.prototype.toObject = function() { 
+    return {
+        "nodes" : this.nodes.toObject(),
+        "result" : this.result.toObject(),
+        "total_result" : this.total_result.toObject(),
+        "task" : this.task.toObject()
+    };
+};
+bm.BenchmarkSummary.prototype.readFromObject = function(json) { 
+    _hasOwnProperty.call(json, "nodes") && (this.nodes.readFromObject(json.nodes));
+    _hasOwnProperty.call(json, "result") && (this.result.readFromObject(json.result));
+    _hasOwnProperty.call(json, "total_result") && (this.total_result.readFromObject(json.total_result));
+    _hasOwnProperty.call(json, "task") && (this.task.readFromObject(json.task));
+    return this;
+};
+bm.BenchmarkSummary.prototype.toBinBuffer = function () {
+    var os = new TarsStream.TarsOutputStream();
+    this._writeTo(os);
+    return os.getBinBuffer();
+};
+bm.BenchmarkSummary.new = function () {
+    return new bm.BenchmarkSummary();
+};
+bm.BenchmarkSummary.create = function (is) {
+    return bm.BenchmarkSummary._readFrom(is);
+};
+
+bm.BenchmarkUnit = function() {
+    this.owner = "";
+    this.servant = "";
+    this.rpcfunc = "";
+    this.para_input = "";
+    this.para_value = "";
+    this.paralist = new TarsStream.List(TarsStream.String);
+    this.endpoints = new TarsStream.List(TarsStream.String);
+    this.links = 0;
+    this.speed = 0;
+    this.duration = 0;
+    this.proto = "";
+    this._classname = "bm.BenchmarkUnit";
+};
+bm.BenchmarkUnit._classname = "bm.BenchmarkUnit";
+bm.BenchmarkUnit._write = function (os, tag, value) { os.writeStruct(tag, value); };
+bm.BenchmarkUnit._read  = function (is, tag, def) { return is.readStruct(tag, true, def); };
+bm.BenchmarkUnit._readFrom = function (is) {
+    var tmp = new bm.BenchmarkUnit;
+    tmp.owner = is.readString(0, false, "");
+    tmp.servant = is.readString(1, false, "");
+    tmp.rpcfunc = is.readString(2, false, "");
+    tmp.para_input = is.readString(3, false, "");
+    tmp.para_value = is.readString(4, false, "");
+    tmp.paralist = is.readList(5, false, TarsStream.List(TarsStream.String));
+    tmp.endpoints = is.readList(6, false, TarsStream.List(TarsStream.String));
+    tmp.links = is.readInt32(7, false, 0);
+    tmp.speed = is.readInt32(8, false, 0);
+    tmp.duration = is.readInt32(9, false, 0);
+    tmp.proto = is.readString(10, false, "");
+    return tmp;
+};
+bm.BenchmarkUnit.prototype._writeTo = function (os) {
+    os.writeString(0, this.owner);
+    os.writeString(1, this.servant);
+    os.writeString(2, this.rpcfunc);
+    os.writeString(3, this.para_input);
+    os.writeString(4, this.para_value);
+    os.writeList(5, this.paralist);
+    os.writeList(6, this.endpoints);
+    os.writeInt32(7, this.links);
+    os.writeInt32(8, this.speed);
+    os.writeInt32(9, this.duration);
+    os.writeString(10, this.proto);
+};
+bm.BenchmarkUnit.prototype._equal = function (anItem) {
+    return this.servant === anItem.servant && 
+        this.rpcfunc === anItem.rpcfunc;
+};
+bm.BenchmarkUnit.prototype._genKey = function () {
+    if (!this._proto_struct_name_) {
+        this._proto_struct_name_ = "STRUCT" + Math.random();
+    }
+    return this._proto_struct_name_;
+};
+bm.BenchmarkUnit.prototype.toObject = function() { 
+    return {
+        "owner" : this.owner,
+        "servant" : this.servant,
+        "rpcfunc" : this.rpcfunc,
+        "para_input" : this.para_input,
+        "para_value" : this.para_value,
+        "paralist" : this.paralist.toObject(),
+        "endpoints" : this.endpoints.toObject(),
+        "links" : this.links,
+        "speed" : this.speed,
+        "duration" : this.duration,
+        "proto" : this.proto
+    };
+};
+bm.BenchmarkUnit.prototype.readFromObject = function(json) { 
+    _hasOwnProperty.call(json, "owner") && (this.owner = json.owner);
+    _hasOwnProperty.call(json, "servant") && (this.servant = json.servant);
+    _hasOwnProperty.call(json, "rpcfunc") && (this.rpcfunc = json.rpcfunc);
+    _hasOwnProperty.call(json, "para_input") && (this.para_input = json.para_input);
+    _hasOwnProperty.call(json, "para_value") && (this.para_value = json.para_value);
+    _hasOwnProperty.call(json, "paralist") && (this.paralist.readFromObject(json.paralist));
+    _hasOwnProperty.call(json, "endpoints") && (this.endpoints.readFromObject(json.endpoints));
+    _hasOwnProperty.call(json, "links") && (this.links = json.links);
+    _hasOwnProperty.call(json, "speed") && (this.speed = json.speed);
+    _hasOwnProperty.call(json, "duration") && (this.duration = json.duration);
+    _hasOwnProperty.call(json, "proto") && (this.proto = json.proto);
+    return this;
+};
+bm.BenchmarkUnit.prototype.toBinBuffer = function () {
+    var os = new TarsStream.TarsOutputStream();
+    this._writeTo(os);
+    return os.getBinBuffer();
+};
+bm.BenchmarkUnit.new = function () {
+    return new bm.BenchmarkUnit();
+};
+bm.BenchmarkUnit.create = function (is) {
+    return bm.BenchmarkUnit._readFrom(is);
+};
+
+var __bm_Admin$query$IF = {
+    "name" : "query",
+    "return" : "int32",
+    "arguments" : [{
+        "name" : "req",
+        "class" : "bm.BenchmarkUnit",
+        "direction" : "in"
+    }, {
+        "name" : "stat",
+        "class" : "bm.ResultStat",
+        "direction" : "out"
+    }]
+};
+
+var __bm_Admin$query$IE = function (req) {
+    var os = new TarsStream.TarsOutputStream();
+    os.writeStruct(1, req);
+    return os.getBinBuffer();
+};
+
+var __bm_Admin$query$ID = function (data) {
+    try {
+        var is = new TarsStream.TarsInputStream(data.response.sBuffer);
+        return {
+            "request" : data.request,
+            "response" : {
+                "costtime" : data.request.costtime,
+                "return" : is.readInt32(0, true, 0),
+                "arguments" : {
+                    "stat" : is.readStruct(2, true, _TARS_MODULE_A_.bm.ResultStat)
+                }
+            }
+        };
+    } catch (e) {
+        throw _makeError(data, e.message, TarsError.CLIENT.DECODE_ERROR);
+    }
+};
+
+var __bm_Admin$query$PE = function (req, __$PROTOCOL$VERSION) {
+    var tup = new TarsStream.UniAttribute();
+    tup.tupVersion = __$PROTOCOL$VERSION;
+    tup.writeStruct("req", req);
+    return tup;
+};
+
+var __bm_Admin$query$PD = function (data) {
+    try {
+        var tup = data.response.tup;
+        return {
+            "request" : data.request,
+            "response" : {
+                "costtime" : data.request.costtime,
+                "return" : tup.readInt32("", 0),
+                "arguments" : {
+                    "stat" : tup.readStruct("stat", _TARS_MODULE_A_.bm.ResultStat)
+                }
+            }
+        };
+    } catch (e) {
+        throw _makeError(data, e.message, TarsError.CLIENT.DECODE_ERROR);
+    }
+};
+
+var __bm_Admin$query$ER = function (data) {
+    throw _makeError(data, "Call Admin::query failed");
+};
+
+bm.AdminProxy.prototype.query = function (req) {
+    var version = this._worker.version;
+    if (version === TarsStream.Tup.TUP_SIMPLE || version === TarsStream.Tup.TUP_COMPLEX) {
+        return this._worker.tup_invoke("query", __bm_Admin$query$PE(req, version), arguments[arguments.length - 1], __bm_Admin$query$IF).then(__bm_Admin$query$PD, __bm_Admin$query$ER);
+    } else {
+        return this._worker.tars_invoke("query", __bm_Admin$query$IE(req), arguments[arguments.length - 1], __bm_Admin$query$IF).then(__bm_Admin$query$ID, __bm_Admin$query$ER);
+    }
+};
+bm.AdminProxy.query = __bm_Admin$query$IF;
+
+var __bm_Admin$shutdown$IF = {
+    "name" : "shutdown",
+    "return" : "int32",
+    "arguments" : [{
+        "name" : "req",
+        "class" : "bm.BenchmarkUnit",
+        "direction" : "in"
+    }, {
+        "name" : "stat",
+        "class" : "bm.ResultStat",
+        "direction" : "out"
+    }]
+};
+
+var __bm_Admin$shutdown$IE = function (req) {
+    var os = new TarsStream.TarsOutputStream();
+    os.writeStruct(1, req);
+    return os.getBinBuffer();
+};
+
+var __bm_Admin$shutdown$ID = function (data) {
+    try {
+        var is = new TarsStream.TarsInputStream(data.response.sBuffer);
+        return {
+            "request" : data.request,
+            "response" : {
+                "costtime" : data.request.costtime,
+                "return" : is.readInt32(0, true, 0),
+                "arguments" : {
+                    "stat" : is.readStruct(2, true, _TARS_MODULE_A_.bm.ResultStat)
+                }
+            }
+        };
+    } catch (e) {
+        throw _makeError(data, e.message, TarsError.CLIENT.DECODE_ERROR);
+    }
+};
+
+var __bm_Admin$shutdown$PE = function (req, __$PROTOCOL$VERSION) {
+    var tup = new TarsStream.UniAttribute();
+    tup.tupVersion = __$PROTOCOL$VERSION;
+    tup.writeStruct("req", req);
+    return tup;
+};
+
+var __bm_Admin$shutdown$PD = function (data) {
+    try {
+        var tup = data.response.tup;
+        return {
+            "request" : data.request,
+            "response" : {
+                "costtime" : data.request.costtime,
+                "return" : tup.readInt32("", 0),
+                "arguments" : {
+                    "stat" : tup.readStruct("stat", _TARS_MODULE_A_.bm.ResultStat)
+                }
+            }
+        };
+    } catch (e) {
+        throw _makeError(data, e.message, TarsError.CLIENT.DECODE_ERROR);
+    }
+};
+
+var __bm_Admin$shutdown$ER = function (data) {
+    throw _makeError(data, "Call Admin::shutdown failed");
+};
+
+bm.AdminProxy.prototype.shutdown = function (req) {
+    var version = this._worker.version;
+    if (version === TarsStream.Tup.TUP_SIMPLE || version === TarsStream.Tup.TUP_COMPLEX) {
+        return this._worker.tup_invoke("shutdown", __bm_Admin$shutdown$PE(req, version), arguments[arguments.length - 1], __bm_Admin$shutdown$IF).then(__bm_Admin$shutdown$PD, __bm_Admin$shutdown$ER);
+    } else {
+        return this._worker.tars_invoke("shutdown", __bm_Admin$shutdown$IE(req), arguments[arguments.length - 1], __bm_Admin$shutdown$IF).then(__bm_Admin$shutdown$ID, __bm_Admin$shutdown$ER);
+    }
+};
+bm.AdminProxy.shutdown = __bm_Admin$shutdown$IF;
+
+var __bm_Admin$startup$IF = {
+    "name" : "startup",
+    "return" : "int32",
+    "arguments" : [{
+        "name" : "req",
+        "class" : "bm.BenchmarkUnit",
+        "direction" : "in"
+    }]
+};
+
+var __bm_Admin$startup$IE = function (req) {
+    var os = new TarsStream.TarsOutputStream();
+    os.writeStruct(1, req);
+    return os.getBinBuffer();
+};
+
+var __bm_Admin$startup$ID = function (data) {
+    try {
+        var is = new TarsStream.TarsInputStream(data.response.sBuffer);
+        return {
+            "request" : data.request,
+            "response" : {
+                "costtime" : data.request.costtime,
+                "return" : is.readInt32(0, true, 0)
+            }
+        };
+    } catch (e) {
+        throw _makeError(data, e.message, TarsError.CLIENT.DECODE_ERROR);
+    }
+};
+
+var __bm_Admin$startup$PE = function (req, __$PROTOCOL$VERSION) {
+    var tup = new TarsStream.UniAttribute();
+    tup.tupVersion = __$PROTOCOL$VERSION;
+    tup.writeStruct("req", req);
+    return tup;
+};
+
+var __bm_Admin$startup$PD = function (data) {
+    try {
+        var tup = data.response.tup;
+        return {
+            "request" : data.request,
+            "response" : {
+                "costtime" : data.request.costtime,
+                "return" : tup.readInt32("", 0)
+            }
+        };
+    } catch (e) {
+        throw _makeError(data, e.message, TarsError.CLIENT.DECODE_ERROR);
+    }
+};
+
+var __bm_Admin$startup$ER = function (data) {
+    throw _makeError(data, "Call Admin::startup failed");
+};
+
+bm.AdminProxy.prototype.startup = function (req) {
+    var version = this._worker.version;
+    if (version === TarsStream.Tup.TUP_SIMPLE || version === TarsStream.Tup.TUP_COMPLEX) {
+        return this._worker.tup_invoke("startup", __bm_Admin$startup$PE(req, version), arguments[arguments.length - 1], __bm_Admin$startup$IF).then(__bm_Admin$startup$PD, __bm_Admin$startup$ER);
+    } else {
+        return this._worker.tars_invoke("startup", __bm_Admin$startup$IE(req), arguments[arguments.length - 1], __bm_Admin$startup$IF).then(__bm_Admin$startup$ID, __bm_Admin$startup$ER);
+    }
+};
+bm.AdminProxy.startup = __bm_Admin$startup$IF;
+
+var __bm_Admin$test$IF = {
+    "name" : "test",
+    "return" : "int32",
+    "arguments" : [{
+        "name" : "req",
+        "class" : "bm.BenchmarkUnit",
+        "direction" : "in"
+    }, {
+        "name" : "rsp",
+        "class" : "string",
+        "direction" : "out"
+    }, {
+        "name" : "errmsg",
+        "class" : "string",
+        "direction" : "out"
+    }]
+};
+
+var __bm_Admin$test$IE = function (req) {
+    var os = new TarsStream.TarsOutputStream();
+    os.writeStruct(1, req);
+    return os.getBinBuffer();
+};
+
+var __bm_Admin$test$ID = function (data) {
+    try {
+        var is = new TarsStream.TarsInputStream(data.response.sBuffer);
+        return {
+            "request" : data.request,
+            "response" : {
+                "costtime" : data.request.costtime,
+                "return" : is.readInt32(0, true, 0),
+                "arguments" : {
+                    "rsp" : is.readString(2, true, ""),
+                    "errmsg" : is.readString(3, true, "")
+                }
+            }
+        };
+    } catch (e) {
+        throw _makeError(data, e.message, TarsError.CLIENT.DECODE_ERROR);
+    }
+};
+
+var __bm_Admin$test$PE = function (req, __$PROTOCOL$VERSION) {
+    var tup = new TarsStream.UniAttribute();
+    tup.tupVersion = __$PROTOCOL$VERSION;
+    tup.writeStruct("req", req);
+    return tup;
+};
+
+var __bm_Admin$test$PD = function (data) {
+    try {
+        var tup = data.response.tup;
+        return {
+            "request" : data.request,
+            "response" : {
+                "costtime" : data.request.costtime,
+                "return" : tup.readInt32("", 0),
+                "arguments" : {
+                    "rsp" : tup.readString("rsp"),
+                    "errmsg" : tup.readString("errmsg")
+                }
+            }
+        };
+    } catch (e) {
+        throw _makeError(data, e.message, TarsError.CLIENT.DECODE_ERROR);
+    }
+};
+
+var __bm_Admin$test$ER = function (data) {
+    throw _makeError(data, "Call Admin::test failed");
+};
+
+bm.AdminProxy.prototype.test = function (req) {
+    var version = this._worker.version;
+    if (version === TarsStream.Tup.TUP_SIMPLE || version === TarsStream.Tup.TUP_COMPLEX) {
+        return this._worker.tup_invoke("test", __bm_Admin$test$PE(req, version), arguments[arguments.length - 1], __bm_Admin$test$IF).then(__bm_Admin$test$PD, __bm_Admin$test$ER);
+    } else {
+        return this._worker.tars_invoke("test", __bm_Admin$test$IE(req), arguments[arguments.length - 1], __bm_Admin$test$IF).then(__bm_Admin$test$ID, __bm_Admin$test$ER);
+    }
+};
+bm.AdminProxy.test = __bm_Admin$test$IF;
+
+
+
