+++ client/package.json
@@ -18,6 +18,7 @@
     "vue": "^2.6.11",
     "vue-cookie": "^1.1.4",
     "vue-i18n": "^7.8.0",
+    "vue-json-component": "^0.4.1",
     "vue-router": "^3.1.5",
     "whatwg-fetch": "^2.0.3"
   },
